Refactor payload creation
